requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pillow==10.2.0
requests==2.31.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pillow==10.2.0
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
beautifulsoup4==4.12.3
black==24.1.0
blinker==1.7.0
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
